scoutfs: update item block and manifest item range
The manifests for level 0 blocks always claimed that they could contain
all keys.  That causes a lot of extra bloom filter lookups when in fact
the blocks contain a very small range of keys.

It's true that we don't know what items a dirty segment is going to
contain, don't want to update the manfiest at every insertion, and have
to find the items in the segments in regular searching.

But when they're finalized we know the items they'll contain and can
update the manifest.  We do that by initializing the item block range to
nonsense and extending it as items are added.  When it's finalized we
update the manifest in memory and in the ring.

Signed-off-by: Zach Brown <zab@versity.com>