scoutfs: avoid stale btree block pointer
The btree walk was storing a pointer to the current btree block that it
was working on.  It later used this when the walk continues and the
block becomes a parent.  But it didn't update this pointer if splitting
changed the block to traverse.  By removing this pointer and using the
block data pointers directly we remove the risk of the pointer going
stale.

Signed-off-by: Zach Brown <zab@versity.com>