scoutfs: add buddy block allocator
Add the block allocator.

Logically we use a buddy allocator that's built from bitmaps for
allocators of each order up to the largest allocator that fits in the
device.  This ends up using two bits per block.

On disk we log modified regions of these bitmaps in chunks in blocks in
a preallocated ring.  We carefully coordinate logging the chunks and the
ring size so that we can always write to the tail of the ring.

There's one allocator and it's only read on mount today.  We'll
eventually have multiple of these allocators covering the device and
nodes will coordinate exclusive access to them.

Signed-off-by: Zach Brown <zab@versity.com>