scoutfs: dirent hashes use linear probing
The current mechanism for dealing with dirent name hash collisions is to
use multiple hash functions.  This won't work great with the btree where
it's expensive to search multiple distant items for a given entry.

Instead of having multiple full precision functions we linearly probe a
given number of hash values after the initial name hash.  Now the slow
colliding path walks adjacent items in the tree instead of bouncing
around the tree.

Signed-off-by: Zach Brown <zab@versity.com>