Add seq field to block map item
The file block mapping code needs to know if an existing block mapping
is dirty in the current transaction or not.  It was doing that by
calling in to the allocator.

Instead of calling in to the allocator we can instead store the seq of
the block in the mapping item.  We also probably want to know the seq of
data blocks to make it possible to discover regions of files that have
changed since a previous seq.

This does increase the size of the block mapping item but they're not
long for this world.  We're going to replace them with proper extent
items in the near future.

Signed-off-by: Zach Brown <zab@versity.com>