Remove unused btree internal WALK_NEXT
In the past the WALK_NEXT enum was used to tell the walking core that
the caller was iterating and that they'd need to advance to sibling
blocks if their key landed off the end of a leaf.  In the current code
that's now handled by giving the walk caller a next_key which will
continue the search from the next leaf.  WALK_NEXT is unused and we can
remove it.

Signed-off-by: Zach Brown <zab@versity.com>