Always initialize btree lockdep class
We have to set the lock class to the btree level to keep
lockdep from building long depdencency chains.  We initialized
allocated blocks for tree growth but not for splitting.  We fix this by
moving the init up into allocation instead of in tree growth.  Now all
the places we get blocks from the block calls are set.

This silences a lockdep warning during merge during rm -rf which is the
first place where multiple blocks in a level are locked.

Signed-off-by: Zach Brown <zab@versity.com>
Reviewed-by: Mark Fasheh <mfasheh@versity.com>