Add scoutfs_block_data_from_contents()
The btree code needs to get a pointer to a whole block from just
pointers to elements that it's sorting.  It had some manual code that
assumed details of the blocks.  Let's give it a real block interface to
do what it wants and make it the block API's problem to figure out how
to do it.

Signed-off-by: Zach Brown <zab@versity.com>
Reviewed-by: Mark Fasheh <mfasheh@versity.com>