Fix dirty item counter calculations
Unfortunately the generic augmented callbacks don't work for our
augmented node bits which specifically reflect the left and right nodes.
We need our own rotation callback and then we have boilerplate for the
other two copy and propagate callbacks.  Once we have to provide
.propagate we can call it instead of our own update_dirty_parents()
equivalent.

In addition some callers messed up marking and clearing dirty.  We only
want to mark dirty item insertions, not all inserted items.  And if we
update an item's keys and values we need to clear and mark it to keep
the counters consistent.

Signed-off-by: Zach Brown <zab@versity.com>