interval-tree: Allow user defined objects as endpoints
Users pass in a comparison function which is used when endpoints need to be
checked against each other. We also put each ITTYPE local definition on it's
own line to facilitate the use of pointers. An upcoming dlm patch will make
use of this to allow for keyed, ranged locking.

Signed-off-by: Mark Fasheh <mfasheh@versity.com>