scoutfs: fix ring first_seq calculation
As we write ring blocks we need to update the first_seq to point at
the first live block in the ring.

The existing calculation gets it wrong and stores the seq of the first
block that we wrote in this commit, not the first ring block that
is still live and would need to be read.

Fix the calculation to so that we set first_seq to the first live block
in the ring.

This fixes the bug where a mount can spin printing the super it's using.
This is the server trying to constantly startup as each server start
fails as it can't read the ring.

Signed-off-by: Zach Brown <zab@versity.com>