scoutfs: remove item reading limit
The item reading limit was intended to minimize latency when we were
directly reading cached manifests.  We're now asking the server to walk
the manifest for us and that's a lot more expensive than querying local
cached blocks.

Let's gulp in an entire segment's worth of items if we can.  We'll have
plenty of opportunity to tune this down later.

Signed-off-by: Zach Brown <zab@versity.com>