scoutfs: add lock caching
We refcount our locks and hold them across system calls. If another node
wants access to a given lock we'll mark it as blocking in the bast and queue
a work item so that the lock can later be released. Otherwise locks are
free'd under memory pressure, unmount or after a timer fires.

Signed-off-by: Mark Fasheh <mfasheh@versity.com>