scoutfs: read items from next segment in level
If the starting key for a segment read doesn't fall in a segment then we
have to include the next segment from that level in the read.  If we
don't then the read can think that there are no more items at that level
and assume that all the items in the upper level are all that exist.

Signed-off-by: Zach Brown <zab@versity.com>