scoutfs: have net.c commit btree blocks
Convert the net server metadata dirtying and committing code to use the
btree instead of the ring.  It has to be careful to setup and teardown
the btree info as it starts up and shuts down the server.

This fixes up some questionable setup/teardown changes made in the
previous patches to convert the manifest and allocator.  We could rebase
the patches to merge those together.  But given that the previous
patches don't work at all without the net updates it might not be worth
the trouble.

Signed-off-by: Zach Brown <zab@versity.com>