scoutfs: allow NULL end around read_items
Let both check_range and read_items take a NULL end.  check_range just
doesn't do anything with the end of the range.  read_items defaults
to trying to read as many items as it can but clamps to the extent of
the segments that intersect with the key.

This will let us incrementally add end arguments to the item functions
that are intially passed in as NULL in callers as we add lock coverage.

Signed-off-by: Zach Brown <zab@versity.com>