scoutfs: lock inode index item iteration
Add locks around inode index item iteration.  This is tricky because the
inode index items are enormous and we can't default to coarse locks that
let it read and iterate over the entire key space.  We use the manifest
to find the next small fixed size region to lock and iterate from.

Signed-off-by: Zach Brown <zab@versity.com>