scoutfs: add cache reading limit to _set_batch
Add an end argument to _set_batch to specify the limit of
items we'll read into the cache.

And it turns out that the loop in _set_batch that meant to cache all the
items covered by the batch didn't try hard enough.  It would stop once
the first key was covered but didn't make sure that the coverage
extended to cover last.  This can happen if segment boundaries happen to
fall within the items that make up the batch.  Fix it up while we're in
here.

Signed-off-by: Zach Brown <zab@versity.com>