scoutfs: don't use delayed downconvert work
The delayed downconvert work wasn't being canceled on shutdown.  60s
after unmount at least the net lock's timer would fire and crash trying
to queue the delayed work on the destroyed workqueue.

Proactively unlocking the locks isn't always beneficial to begin with.
The relative costs of mispredicting the future are wildly different if
we have to re-read item caches from segments or have to downconvert a
blocking read lock.

So we can just remove the delayed work to fix the bug and remove a
moving piece that would need to be considered and tuned.  There's still
a race where we can get basts after destroying the workqueue but before
we destroy the lockspace, we'll get there.

Signed-off-by: Zach Brown <zab@versity.com>