scoutfs: add global lock names
Add a lock name that has a global scope in a given lockspace.  It's not
associated with any file system items.  We add a scope to the lock name
to indicate if a lock is global or not and set that in other lock naming
intitialization.  We permit lock allocation to accept null start and end
keys.

Signed-off-by: Zach Brown <zab@versity.com>