scoutfs: delete inode index items
Delete inode index items when deleting all the items associated with an
inode after its been unlinked and had all its references dropped.

The index items should always match the fields in the inode item so we
read it to determine the index items that should be deleted, regardless
of if we have the vfs inode cached or not.  We take the opportunity to
collapse the two callers of item deletion which looked up the inode into
item deletion so that it can use the inode fields.

The deletion of index items is partially verified by an inode index test
in xfstests which makes sure that unlinked files are no longer present
in the index.

Signed-off-by: Zach Brown <zab@versity.com>