scoutfs: give module fs scoutfs alias
Use MODULE_ALIAS_FS() to register the "scoutfs" fs alias so that
modprobe can find the module if it's installed and visible to depmod.

We don't yet have clever enough xfstests to mess around with modules.  I
manually verified this by installing the module in /lib/modules and
trying mount -t scoutfs before and after the change.

Signed-off-by: Zach Brown <zab@versity.com>