scoutfs: print trailing key bytes
The key printing functions only output the key material that's described
by the format.  We have some callers that need to increment or decrement
keys so they expand them to full size keys.  This expansion and extra
high precision low significance was hidden from the traces.

This adds a helper that prints the key material with the format and then
appends an encoding of the trailing bytes.

The key printer was a huge mess of cases and ifs that made it hard to
integrate a sane helper.  We also take the opportunity to break it up
into zone|type key printer functions.  The isolation makes it much
clearer to see what's going on.

Signed-off-by: Zach Brown <zab@versity.com>