scoutfs: dlmglue should initialize res->l_debug_list
We're missing initialization of this field. It should never cause a
problem today because we always do a list add immediately afterwards
but let's be extra careful here and initialize it just in case. We
also add a sanity check in ocfs2_add_lockres_tracking() that the
lockres hasn't already been put on the debug list.

Signed-off-by: Mark Fasheh <mfasheh@versity.com>