scoutfs: warn on invalid item counts
We had a bug where a caller was slowly increasing their item count for
every transaction they attempted in a loop.  Eventually the item count
grew to be too large to fit in a segment and they slept indefinitely.

Let's warn on invalid and impossibly large item counts as we enter
transactions.

Signed-off-by: Zach Brown <zab@versity.com>