scoutfs: add item count update function
The item cache maintains a count of the number of dirty items, keys, and
values.  It updates the counts as it dirties and cleans items.  There
are callers who want to modify the accounting directly instead of having
the accounting updated as a side effect of cleaning and re-dirtying the
item.

Signed-off-by: Zach Brown <zab@versity.com>