scoutfs: fix btree bit iteration
store_pos_bits() was trying to iterate over bits that were different
between the existing bits set in the item and the new bits that will be
set.  It used a too clever for_each helper that tried to only iterate as
many times as there were bits.  But it messed up and only used ffs to
find the next bit for the first iteration.  From then on it would
iterate over bits that weren't different.  This would cause the counts
to be changed when the bits didn't change and end up being wildly wrong.

Fix this by using a much clearer loop.  It still breaks out when there
are no more different bits and we're only using a few low bits so the
number of iterations is tiny.

Signed-off-by: Zach Brown <zab@versity.com>