scoutfs: more reliably set btree parent item bits
Most paths correctly calculated the bits to set in a parent item by
combining the bits set in the child with the half bit of the child
block's position in the ring.

With the exception of fixing up the parent item bits after descent by
walking the path.  This mistake caused the parent item half bits to be
zero and prevented migrating of blocks from the old half of the ring.

Fix it by introducing a helper function to calculate the parent ref item
bits and consistently using it.

Signed-off-by: Zach Brown <zab@versity.com>