scoutfs: dlmglue/lock counters and tracing
We can use the excellent code in counters.h to easily place a whole set
of useful counters in dlmglue:

- one for every kind of wait in cluster_lock (blocked, busy, etc)
- one for each type of dlm operation (lock/unlock requests,
  converts, etc)
- one for each type of downconvert (cw/pr/ex)

These will give us a decent idea of the amount and type of lock traffic a
given node is seeing.

In addition, we add a second trace at the bottom of invalidate_caches.
By turning both traces in invalidate_caches on, we can look at our
trace log to see how long a given locks downconvert took.

Signed-off-by: Mark Fasheh <mfasheh@versity.com>