scoutfs: add trigger to drop lock cache
We have a corruption that can happen when a lock is reclaimed but it's
cache is still dirty. Detect this corruption by placing a trigger in
statfs which fires off lock reclaim. Statfs is nice because for scoutfs
it's lockless, which means there should not be any references on locks
when the trigger is fired.

Signed-off-by: Mark Fasheh <mfasheh@versity.com>