scoutfs: add debug check for scout-107
We have a bug filed where the fs got stuck spinning in
scoutfs_dir_get_backref_path(). There's been enough changes lately that
we're not sure if this issue still exists. Catch if we have an excessive
number of iterations through our loop there and exit with some debug info.

Signed-off-by: Mark Fasheh <mfasheh@versity.com>