scoutfs: track inode 512b block count
We weren't doing anything with the inode blocks field.  We weren't even
initializing it which explains why we'd sometimes see garbage i_blocks
values in scoutfs inodes in segments.

The logical blocks field reflects the contents of the file regardless of
whether its online or not.  It's the sum of our online and offline block
tracking.

So we can initialize it to our persistent online and offline counts and
then keep it in sync as blocks are allocated and freed.

Signed-off-by: Zach Brown <zab@versity.com>