scoutfs: reserve at least one xattr item value
Even when we're setting an xattr with no value we still have a file
system item value that contains the xattr value header which tells us
that this is the last value.

This fixes a warning that would be issued if we tried to set an xattr
with a zero length value.  We'd try to dirty an item value with the
header after having reserved zero bytes for item values.  To hit the
warning the inode couldn't already be dirty so that the xattr value
didn't get to hide in the unsed reservation for dirtying the inode
item's value.

Signed-off-by: Zach Brown <zab@versity.com>