scoutfs: add some lock/item consistency checks
Add some tests to the locking paths to see if we violate item caching
rules.

As we finish locking calls we make sure that the item cache is
consistent with the lock mode.  And we make sure that we don't free
locks before they've been unlocked and had a chance to check the
item cache.

Signed-off-by: Zach Brown <zab@versity.com>