scoutfs: avoid double unlock
We weren't sufficiently careful in reacting to basts.  If a bast arrived
whlie an unlock is in flight we'd turn around and try to unlock again,
returning an error, and exploding.

More carefully only act on basts if we have an active mode that needs to
be unlocked.  Now if the racey bast arrives we'll ignore it and end up
freeing the lock in processing after the unlock succeeds.

Signed-off-by: Zach Brown <zab@versity.com>