scoutfs: alloc inos at multiple of lock group
Inode allocations come from batches that are reserved for directories.
As the batch is exhausted a new one is acquired and allocated from.

The batch size was arbitrarily set to the human friendly 10000.  This
doesn't interact well with the lock group size being a power of two.
Each allocation batch will straddle an inode group with its previous and
next inode batch.

This often doesn't matter because dirctories very rarely have more than
9000 entries.  But as entries pass 10000 they'd see surprising
contention with other inode ranges in directories.

Tweak the allocation size to be a multiple of the lock group size to
stop this from happening.

Signed-off-by: Zach Brown <zab@versity.com>