scoutfs: store file data in extents
Store file data mappings and free block ranges in extents instead of in
block mapping items and bitmaps.

This adds the new functionality and refactors the functions that use it.
The old functions are no longer called and we stop at ifdeffing them out
to keep the change small.  We'll remove all the dead code in a future
change.

Signed-off-by: Zach Brown <zab@versity.com>