scoutfs: add scoutfs_lock_rename()
Add a specific lock method for locking the global rename lock instead of
having the caller specify it as a global lock.  We're getting rid of the
notion of lock scopes and requiring all locks to be related to keys.
The rename lock will use magic keys at the end of the volume.

Signed-off-by: Zach Brown <zab@versity.com>