scoutfs: memmove deleted btree items
It turns out that the sorting performed by btree block item compaction
was pretty expensive.  It's cheaper to keep the items packed at the end
of the block by moving earlier items towards the back of the block as
interior items are deleted.  When the items are always packed at the end
of the block we no longer need to track fragmented free space and can
remove the 'free_reclaim' btree block field.

This brought the bulk empty file create rate up by about 20%.

Signed-off-by: Zach Brown <zab@versity.com>